api/auth/requirements.txt
alembic==1.4.3
attrs==20.2.0
awscli==1.18.107
awsebcli==3.19.0
bcrypt==3.2.0
blessed==1.17.10
blinker==1.4
boto3==1.14.30
botocore==1.17.30
cached-property==1.5.1
cachetools==4.1.1
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.1
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
flake8==3.8.3
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Shopify==0.2
Flask-SQLAlchemy==2.4.4
future==0.16.0
google-auth==1.21.2
google-auth-oauthlib==0.4.1
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.7
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.4
paramiko==2.7.2
pathspec==0.5.9
pluggy==0.13.1
py==1.9.0
pyactiveresource==2.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.0.1
python-dateutil==2.8.0
python-dotenv==0.14.0
python-editor==1.0.4
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.3
semantic-version==2.5.0
ShopifyAPI==8.0.1
six==1.11.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
stripe==2.50.0
termcolor==1.1.0
texttable==1.6.3
toml==0.10.1
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.1.0
